[NO-ISSUE] refactor: 패키지 변경 (utils -> util)
