+++ src/main/java/com/sds/actlongs/util/manage/file/FileManage.java
@@ -1,4 +1,4 @@
-package com.sds.actlongs.utils.manage.file;
+package com.sds.actlongs.util.manage.file;
 
 import java.io.InputStream;
 import java.nio.file.Path;
