+++ src/main/java/com/sds/actlongs/util/uuid/UuidGenerate.java
@@ -1,4 +1,4 @@
-package com.sds.actlongs.utils.uuid;
+package com.sds.actlongs.util.uuid;
 
 public interface UuidGenerate {
 	String getVodUuid();
