+++ src/main/java/com/sds/actlongs/util/uuid/UuidGenerateImpl.java
@@ -1,4 +1,4 @@
-package com.sds.actlongs.utils.uuid;
+package com.sds.actlongs.util.uuid;
 
 import java.util.UUID;
